refactor(ingest): split apply.py engine into a stage package
Turn the 1312-line apply.py into an apply/ package, one module per
pipeline stage, dependencies flowing one way (leaves -> composers):

- validate.py   structural plan checks (leaf)
- claims.py     build -> validate-values -> diff; owns RetractEntry (leaf)
- dry_run.py    read-only validate + diff; depends on claims
- persist.py    entity creation, citations, claim/ChangeSet writes,
                resolve; depends on claims; holds the two catalog
                touchpoints (build_relationship_claim, resolve_all_entities)
- orchestrate.py  apply_plan: transaction structure + run bookkeeping
- __init__.py   re-exports apply_plan (public import path unchanged)

Pure code motion, no behavior change. Drops the dead module-level
logger. Public surface stays exactly apply_plan, imported from the
package; no caller changes.
